Move files into own classes
